Merge branch 'master' of git.linux.iastate.edu:309Spring2016/G07_Flock into CK-2
Finished putting all elements of an event on a page.  It does not look like the final product, but all elements are there so I can continue onto the next part I will be working on.  I have also created some Unit Tests that test both of the controllers I have created.